+++ amundsen_application/static/js/pages/ProfilePage/index.tsx
@@ -8,10 +8,8 @@ import { connect } from 'react-redux';
 import { RouteComponentProps, withRouter } from 'react-router';
 import { bindActionCreators } from 'redux';
 
-import Breadcrumb from 'components/common/Breadcrumb';
-import Flag from 'components/common/Flag';
+import Breadcrumb from 'components/Breadcrumb';
 import TabsComponent, { TabInfo } from 'components/common/TabsComponent';
-import { BadgeStyle } from 'config/config-types';
 
 import { GlobalState } from 'ducks/rootReducer';
 import { getUser, getUserOwn, getUserRead } from 'ducks/user/reducer';
