+++ amundsen_application/static/js/pages/TableDetailPage/index.tsx
@@ -29,7 +29,7 @@ import {
 
 import BadgeList from 'features/BadgeList';
 import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
-import Breadcrumb from 'components/common/Breadcrumb';
+import Breadcrumb from 'components/Breadcrumb';
 import TabsComponent, { TabInfo } from 'components/common/TabsComponent';
 import TagInput from 'components/common/Tags/TagInput';
 import EditableText from 'components/common/EditableText';
